Move InfoDialog main class to infodialog/dialog.py
